% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Bar Chart}
\usage{
bar_chart(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  bar_colours = orr_colours(),
  y_axis_breaks = ggplot2::waiver(),
  y_axis_labeller = scales::label_comma(),
  data_labeller = label_orr_comma(),
  show_legend = TRUE
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{bar_colours}{An array of colour hex-codes. Defaults to ORR colours.}

\item{y_axis_breaks}{Vector of where to place y-axis breaks. If blank will
use ggplot2 default.}

\item{y_axis_labeller}{A function which creates the y-axis label strings.}

\item{data_labeller}{A function which controls how the data labels over the bars are displayed.}

\item{show_legend}{Should the legend be shown on the chart. Defaults to \code{TRUE}.}
}
\description{
A simple or stacked bar chart in ORR style.
}
