% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_report.R
\name{board_report_gb_line_chart}
\alias{board_report_gb_line_chart}
\title{Board Report GB Line Chart}
\usage{
board_report_gb_line_chart(
  data,
  filename,
  path = NULL,
  gb_target_value,
  chart_width = 2.6575,
  chart_height = 1.6457,
  y_axis_breaks = ggplot2::waiver(),
  y_axis_labeller = scales::label_comma(),
  last_point_labeller = scales::label_comma()
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{gb_target_value}{The value of the end of year target for the GB metric.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{y_axis_breaks}{Vector of where to place y-axis breaks. If blank will
use ggplot2 default.}

\item{y_axis_labeller}{A function which creates the y-axis label strings.}

\item{last_point_labeller}{A function which determines how data labels are displayed.}
}
\description{
Create a line chart for the board report with GB values in black and regions in grey.
}
