% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_report.R
\name{board_report_region_bars}
\alias{board_report_region_bars}
\title{Board Report Region Bars}
\usage{
board_report_region_bars(
  data,
  filename,
  path = NULL,
  region_targets,
  period_number,
  metric_name,
  chart_width = 2.6575,
  chart_height = 1.6457,
  bar_colour = "#253268",
  point_colour = "#D8730F",
  data_labeller = label_orr_comma()
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{region_targets}{A dataframe with two columns containing the region name and target value.}

\item{period_number}{The current period number, shown in the subtitle.}

\item{metric_name}{The name of the metric plotted, shown in the chart title.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{bar_colour, point_colour}{Set the colours of bars and points in the plot.}
}
\description{
Create a board report plot with bars and points for targets.
}
