% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_to_orr.R
\name{ggplot_to_orr}
\alias{ggplot_to_orr}
\title{Ggplot2 To ORR}
\usage{
ggplot_to_orr(
  plot,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567
)
}
\arguments{
\item{plot}{A ggplot2 plot object.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}
}
\description{
Take a custom ggplot and apply basic ORR theme and styling, then save as a
PNG file. Try to use the chart functions where ever possible but this is
sometimes helpful for a bespoke chart.
}
