% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_side_by_side_bar.R
\name{grouped_side_by_side_bar}
\alias{grouped_side_by_side_bar}
\title{Grouped Side-By-Side Bar Chart}
\usage{
grouped_side_by_side_bar(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  left_bar_title = "METRIC, TIME PERIOD",
  right_bar_title = "Change from TIME PERIOD",
  left_bar_colour = "#253268",
  right_bar_colour = "#D8730F",
  left_bar_labeller = label_orr_comma(),
  right_bar_labeller = label_orr_percent(),
  text_outside_bar_threshold = 0.2,
  rightside_margin = NA
)
}
\arguments{
\item{data}{A dataframe with 4 columns. The first column should hold the bar categories, which set the y-axis labels (not in bold). The second column hold the groups of the bar categories, shown on the y-axis labels in bold. Column 3 holds the values for the left had set of bars. Column 4 holds the values for the right hand set of bars.The vertical order of the bars within the group matches the order they are supplied in the dataframe.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output.}

\item{left_bar_title, right_bar_title}{The text which appears above each set
of bars.}

\item{left_bar_colour}{A string hexcode for the colour of the bars in the
left-hand side bar chart. Default is for ORR dark blue.}

\item{right_bar_colour}{String hexcodes for the right hand set of bars.}

\item{left_bar_labeller, right_bar_labeller}{Functions, for example result of
a labeller function, which turn data numeric values into character labels.
See \link[scales]{label_percent}}

\item{text_outside_bar_threshold}{Proportion of largest bar at which data labels are shown outside the bar. Default is 0.2.}

\item{rightside_margin}{Expand the margin on the right-hand side of the plot if a data label is cut off.}
}
\description{
Two columns of horizontal bars, grouped with titles.
}
