% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_orr_comma}
\alias{label_orr_comma}
\title{Label ORR Comma}
\usage{
label_orr_comma(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector. Numbers are shown with comma every 3 digits and
rounded to 3 significant figures.
}
\description{
Turn large numbers into string in ORR format.
}
\seealso{
\code{\link[=label_orr_percent]{label_orr_percent()}}, \code{\link[=label_orr_percentage_point]{label_orr_percentage_point()}}
}
