% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_orr_percent}
\alias{label_orr_percent}
\title{Label ORR Percent}
\usage{
label_orr_percent()
}
\value{
A labeller function. Values are shown with a '\%' symbol to zero
decimal places unless between 1\% and -1\% and then shown to 1 decimal place.
}
\description{
Turn number into an string in ORR percent format.
}
\seealso{
\code{\link[=label_orr_comma]{label_orr_comma()}}, \code{\link[=label_orr_percentage_point]{label_orr_percentage_point()}}
}
