% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orrcharts-package.R
\docType{package}
\name{orrcharts-package}
\alias{orrcharts}
\alias{orrcharts-package}
\title{orrcharts: Charts For ORR Stats Reports}
\description{
Functions to support the production of ORR stats releases.
}
\author{
\strong{Maintainer}: Oli Elliott \email{oliver.elliott@orr.gov.uk}

}
\keyword{internal}
