% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly_bar.R
\name{quarterly_bar}
\alias{quarterly_bar}
\title{Quarterly Bar Chart}
\usage{
quarterly_bar(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  bar_colour = "#253268",
  v_nudge_data_label = 0,
  h_nudge_x_axis_labels = 0,
  y_axis_breaks = ggplot2::waiver(),
  y_axis_labeller = scales::label_percent(scale = 1, accuracy = 1),
  last_point_labeller = scales::label_percent(scale = 1, accuracy = 0.1)
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{bar_colour}{A hex code or colour name for the fill colour of the bars.
Defaults to ORR dark blue.}

\item{v_nudge_data_label}{Shift the last bar data label up or down to dodge
the neighbouring bars. Negative values move upwards, positive values move
downwards.}

\item{h_nudge_x_axis_labels}{Shift the x-axis labels right or left to be in
the centre of the years. Typically between -1 and 1. Negative value shifts
right, positive shifts left.}

\item{y_axis_breaks}{Vector of where to place y-axis breaks. If blank will
use ggplot2 default.}

\item{y_axis_labeller}{A function which creates the y-axis label strings.}

\item{last_point_labeller}{A function which create the data label over the
last point from the value.}
}
\value{
A ggplot2 bar chart for ORR quarterly bar time series.
}
\description{
A bar chart time series with the label above the last data point.
}
