% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_by_side_bar.R
\name{side_by_side_bar}
\alias{side_by_side_bar}
\title{Side-By-Side Bar Chart}
\usage{
side_by_side_bar(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 7.3,
  left_bar_title = "METRIC, TIME PERIOD",
  right_bar_title = "Change from TIME PERIOD",
  order_by_bar = c("right", "left"),
  order_descending = TRUE,
  left_bar_colour = "#253268",
  right_bar_colour_positive = ifelse(order_descending, "#28994b", "#B1173B"),
  right_bar_colour_negative = ifelse(order_descending, "#B1173B", "#28994b"),
  left_bar_labeller = label_orr_comma(),
  right_bar_labeller = label_orr_percent(),
  panel_proportional_widths = c(0.365, 0.28),
  panel_axes_expansion = c(0.2, 0.075),
  right_bar_order_descending = order_descending
)
}
\arguments{
\item{data}{A data frame with the first 3 columns holding the TOC names or
y-axis category, the left bar value, the right bar change value.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output.}

\item{left_bar_title, right_bar_title}{The text which appears above each set
of bars.}

\item{order_by_bar}{Order by the left or right set of bars.}

\item{order_descending}{Sort bars descending or ascending.}

\item{left_bar_colour}{A string hexcode for the colour of the bars in the
left-hand side bar chart. Default is for ORR dark blue.}

\item{right_bar_colour_positive, right_bar_colour_negative}{String hexcodes
for the colours of the right-hand side bar chart. \verb{_positive} colours
positive valued bars and \verb{_negative} colours those below zero. Default is
for green/red but could also use ORR yellow for both.}

\item{left_bar_labeller, right_bar_labeller}{Functions, for example result of
a labeller function, which turn data numeric values into character labels.
See \link[scales]{label_percent}}

\item{panel_proportional_widths}{Set the widths of the left and right plots
in proportion to \code{chart_width}. Space needs to be left for the y-axis
labels.}

\item{panel_axes_expansion}{Multiplicative axis expansion values for the
bars. Increase the first value to make more space for data labels. Increase
the 2nd value to make more space between the charts.}

\item{right_bar_order_descending}{This parameter is deprecated to allow
sorting by left or right bar. \code{TRUE} or \code{FALSE}. If \code{TRUE} both bar charts
are arranged in descending order of the right-hand bar chart. If \code{FALSE}
then ascending and default bar colours flip.}
}
\description{
Side-by-side bar chart to compare changes over time in a metric between TOCs.
}
