% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_chart.R
\name{donut_chart}
\alias{donut_chart}
\title{Donut Chart}
\usage{
donut_chart(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  colours = orr_colours(),
  data_labeller = scales::label_number(scale = 1, accuracy = 1),
  labels_gap_size = 2,
  nudge_label_out = NULL,
  nudge_label_clockwise = NULL,
  min_label_segment_length = 0.4,
  outer_chart_limit = 7,
  as_pie_chart = FALSE,
  centre_label = "",
  seed = 101
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{colours}{The colours for the donut sections. Defaults to ORR colours.}

\item{data_labeller}{A function which controls how the data labels over the
bars are displayed.}

\item{labels_gap_size}{Gap between the outside of the donut and the centre of
the text labels.Labels are positions using ggrepel to avoid overlaps. This
parameter sets the starting distance before the repel algorithm is run.}

\item{nudge_label_out}{An array the same length as rows in data. Positive
values move single labels further out from the \code{labels_gap_size} value.}

\item{nudge_label_clockwise}{An array the same length as rows in data.
Positive values move labels a percentage of 360 degrees clockwise from
their starting positions mid segment. Suggest values between +1 and -1.}

\item{min_label_segment_length}{Minimum length of line between the text label
and the edge of the segment. Hidden if smaller than this.}

\item{outer_chart_limit}{The upper limit on x-axis which sets that outer size
of the chart.}

\item{as_pie_chart}{If true will remove hole in donut and show as pie chart.}

\item{centre_label}{Text displayed at the centre of the chart}

\item{seed}{Random seed for the label positioning algorithm}
}
\description{
A simple donut chart.
}
