% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_chart.R
\name{donut_chart}
\alias{donut_chart}
\title{Donut Chart}
\usage{
donut_chart(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  colours = orr_colours(),
  data_labeller = scales::label_number(scale = 1, accuracy = 1),
  labels_gap_size = 1,
  as_pie_chart = FALSE
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{colours}{The colours for the donut sections. Defaults to ORR colours.}

\item{data_labeller}{A function which controls how the data labels over the bars are displayed.}

\item{labels_gap_size}{Gap between the outside of the donut and the centre of
the text labels.}

\item{as_pie_chart}{If true will remove hole in donut and show as pie chart.}
}
\description{
A simple donut chart.
}
