% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_bar.R
\name{horizontal_bar}
\alias{horizontal_bar}
\title{Horizontal Bar Chart}
\usage{
horizontal_bar(
  data,
  filename,
  path = NULL,
  chart_width = 6.7,
  chart_height = 3.567,
  data_labeller = scales::label_number(),
  bar_colours = orr_colours(),
  outside_bar_threshold = 0.1,
  outside_bar_distance = 0.05,
  show_legend = FALSE,
  legend_position = c(0.8, 0.2)
)
}
\arguments{
\item{data}{A data frame with 2 columns. The first column should be the
financial_quarter_key and the second column the value to be plotted. The
data should start at Q1 of a year.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{data_labeller}{A function which controls how the data labels over the
bars are displayed.}

\item{bar_colours}{An array of colour hex-codes. Defaults to ORR colours.}

\item{outside_bar_threshold}{Controls when a data label is placed inside or
outside the bar. Increase the value if data labels look cramped in small
bars.}

\item{outside_bar_distance}{Controls how far outside the bar labels are
shown. Increase to move further from the bar.}

\item{show_legend}{Should the legend be shown on the chart. Defaults to
\code{TRUE}.}

\item{legend_position}{Set location of the legend in the plot. Default is
bottom right. c(0,0) is bottom left.}
}
\description{
A simple horizontal bar chart to display a single set of values
}
