% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_orr_percentage_point}
\alias{label_orr_percentage_point}
\title{Label ORR Percentage Point Change}
\usage{
label_orr_percentage_point(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.Numbers are shown to 1 decimal place with a 'pp'
suffix and +/- prefix.
}
\description{
Turn number into string in ORR percentage point change format.
}
\seealso{
\code{\link[=label_orr_percent]{label_orr_percent()}}, \code{\link[=label_orr_comma]{label_orr_comma()}}
}
