% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  data,
  filename,
  path,
  chart_width = 6.7,
  chart_height = 3.567,
  series_colours = orr_colours(),
  point_shapes = c(15, 16, 17, 18, 0, 1, 2, 5),
  x_axis_labels = NULL,
  y_axis_breaks = ggplot2::waiver(),
  y_axis_labeller = scales::label_comma(),
  show_series_labels = TRUE,
  data_labeller = label_orr_comma(),
  chart_seed = 101
)
}
\arguments{
\item{data}{A data frame with at least 2 columns. The first column sets the
x-axis, typically a date or time period, which need to be ordered oldest to
newest. Each column after that contains the values to be plotted in each
line. The column names will form the series labels shown on the chart.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{chart_width, chart_height}{Width and height in inches of the PNG output}

\item{series_colours}{Colours for the lines. Defaults to ORR colours.}

\item{point_shapes}{The shapes to show on data point per series. See
\code{\link[ggplot2:scale_shape]{ggplot2::scale_shape()}}.}

\item{x_axis_labels}{An array of labels to display on the x-axis. These need
to match the data supplied in the first column. X-axis ticks will show for
all points but this parameter control which labels are displayed. By
default all labels are shown.}

\item{y_axis_breaks}{Vector of where to place y-axis breaks. If blank will
use ggplot2 default.}

\item{y_axis_labeller}{A function which creates the y-axis label strings.}

\item{show_series_labels}{If \code{TRUE} chart will display the series labels,
taken from column names, on the chart. Label colours will match series
colours.}

\item{data_labeller}{Function which controls how the data labels are
displayed.}

\item{chart_seed}{Set random seed for \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}. \code{ggrepel}
uses random numbers in the positioning algorithm. By setting a seed for the
random number generator the labels will be in the same place each time the
plot is created with the same data.}
}
\value{
A PNG file of a simple ORR style line chart. This function uses an
algorithm from \code{ggrepel} to place the data and series labels. The algorithm
moves the labels in small increments at random trying to minimise the
amount of overlaps. If you are not happy with the final positions, change
the \code{chart_seed} argument - this might not always return different results.
}
\description{
A simple line chart to plot time series data.
}
